Index

This page lists the core functions and types of the Julia component of the Rocketeer.jl package.

"Table" of Contents

Modules	Functions	Types	Constants

Index	Index	Index	Index

Docs	Docs	Docs	Docs

Index

This section enumerates the names exported by the package, each of which links to its corresponding Documentation.

Modules

Rocketeer.Rocketeer

Functions

Rocketeer.apply_kernel

Rocketeer.apply_kernels

Rocketeer.load_rocket

Rocketeer.save_rocket

Types

Rocketeer.RocketKernel

Rocketeer.RocketModule

Rocketeer.RocketModule

Rocketeer.RocketModule

Constants

Rocketeer.ROCKETEER_VERSION

Docs

This section lists the documentation for every exported name of the Rocketeer package.

Modules

Rocketeer.Rocketeer — Module

Main module for the Rocketeer.jl package.

Attribution

Programmer

Sasha Petrenko <petrenkos@mst.edu> @AP6YC

Original Authors

Angus Dempster

Francois Petitjean

Geoff Webb

Bibtex Entry

@article{dempster_etal_2020,
    author  = {Dempster, Angus and Petitjean, Francois and Webb, Geoffrey I},
    title   = {ROCKET: Exceptionally fast and accurate time classification using random convolutional kernels},
    year    = {2020},
    journal = {Data Mining and Knowledge Discovery},
    doi     = {https://doi.org/10.1007/s10618-020-00701-z}
}

Citation Links

Papers:

Journal article (DOI)

Preprint

Software

rocket (Python)

Imports

The following names are imported by the package as dependencies:

Base

Core

DocStringExtensions

NumericalTypeAliases

Pkg

Random

Exports

The following names are exported and available when using the package:

ROCKETEER_VERSION

RocketKernel

RocketModule

apply_kernel

apply_kernels

load_rocket

save_rocket

Functions

Rocketeer.apply_kernel — Method

apply_kernel(
    kernel::RocketKernel,
    x::AbstractVector{T} where T<:Real
) -> Any

Summary

Apply a single RocketKernel to the sequence x.

Arguments

kernel::RocketKernel: the RocketKernel used for computing features.

x::RealVector: data sequence for computing Rocket features.

Method List / Definition Locations

apply_kernel(kernel, x)

defined at /home/runner/work/Rocketeer.jl/Rocketeer.jl/src/lib/rocket.jl:136.

Rocketeer.apply_kernels — Method

apply_kernels(
    rocket::RocketModule,
    x::AbstractVector{T} where T<:Real
) -> Matrix{Float64}

Summary

Run a vector of RocketKernels along a sequence x.

Arguments

rocket::RocketModule: rocket module containing many kernels for processing.

x::RealVector: data sequence for computing rocket features.

Method List / Definition Locations

apply_kernels(rocket, x)

defined at /home/runner/work/Rocketeer.jl/Rocketeer.jl/src/lib/rocket.jl:173.

Rocketeer.load_rocket — Function

load_rocket() -> Any
load_rocket(filepath::AbstractString) -> Any

Summary

Load and return a RocketModule with existing parameters from a .jld2 file.

Arguments

filepath::AbstractString: default rocket.jld2, path to the .jld2 containing rocket parameters.

Method List / Definition Locations

load_rocket()
load_rocket(filepath)

defined at /home/runner/work/Rocketeer.jl/Rocketeer.jl/src/lib/rocket.jl:210.

Rocketeer.save_rocket — Function

save_rocket(rocket::RocketModule)
save_rocket(rocket::RocketModule, filepath::AbstractString)

Summary

Save the RocketModule parameters to a .jld2 file.

Arguments

rocket::RocketModule: the RocketModule to save.

filepath::AbstractString: default rocket.jld2, path to .jld2 for saving rocket parameters.

Method List / Definition Locations

save_rocket(rocket)
save_rocket(rocket, filepath)

defined at /home/runner/work/Rocketeer.jl/Rocketeer.jl/src/lib/rocket.jl:196.

Types

Rocketeer.RocketKernel — Type

struct RocketKernel

Summary

Structure containing information about one Rocket kernel.

Attribution

Programmer

Sasha Petrenko <petrenkos@mst.edu> @AP6YC

Original Authors

Angus Dempster

Francois Petitjean

Geoff Webb

Bibtex Entry

@article{dempster_etal_2020,
    author  = {Dempster, Angus and Petitjean, Francois and Webb, Geoffrey I},
    title   = {ROCKET: Exceptionally fast and accurate time classification using random convolutional kernels},
    year    = {2020},
    journal = {Data Mining and Knowledge Discovery},
    doi     = {https://doi.org/10.1007/s10618-020-00701-z}
}

Citation Links

Papers:

Journal article (DOI)

Preprint

Software

rocket (Python)

Fields

length::Int64: The length of the kernel.

weight::Vector{Float64}: The vector of weights corresponding to the features.

bias::Float64: The internal Rocket bias parameter, computed during construction.

dilation::Int64: The internal Rocket dilation parameter, computed during construction.

padding::Int64: The internal Rocket padding parameter, computed during construction.

Rocketeer.RocketModule — Type

struct RocketModule

Summary

Structure containing a vector of RocketKernels.

Attribution

Programmer

Sasha Petrenko <petrenkos@mst.edu> @AP6YC

Original Authors

Angus Dempster

Francois Petitjean

Geoff Webb

Bibtex Entry

@article{dempster_etal_2020,
    author  = {Dempster, Angus and Petitjean, Francois and Webb, Geoffrey I},
    title   = {ROCKET: Exceptionally fast and accurate time classification using random convolutional kernels},
    year    = {2020},
    journal = {Data Mining and Knowledge Discovery},
    doi     = {https://doi.org/10.1007/s10618-020-00701-z}
}

Citation Links

Papers:

Journal article (DOI)

Preprint

Software

rocket (Python)

Fields

input_length::Int64: The input length used to generate the RocketKernels.

kernels::Vector{RocketKernel}: The list of RocketKernels constituting a full Rocket module.

Rocketeer.RocketModule — Method

RocketModule(
    input_length::Integer,
    n_kernels::Integer
) -> RocketModule

Summary

Create a new RocketModule structure, requiring feature length and the number of kernels.

Arguments

input_length::Integer: the desired length of the kernel features.

n_kernels::Integer: the desired number of kernels to generate.

Attribution

Programmer

Sasha Petrenko <petrenkos@mst.edu> @AP6YC

Original Authors

Angus Dempster

Francois Petitjean

Geoff Webb

Bibtex Entry

@article{dempster_etal_2020,
    author  = {Dempster, Angus and Petitjean, Francois and Webb, Geoffrey I},
    title   = {ROCKET: Exceptionally fast and accurate time classification using random convolutional kernels},
    year    = {2020},
    journal = {Data Mining and Knowledge Discovery},
    doi     = {https://doi.org/10.1007/s10618-020-00701-z}
}

Citation Links

Papers:

Journal article (DOI)

Preprint

Software

rocket (Python)

Method List / Definition Locations

RocketModule(input_length, n_kernels)

defined at /home/runner/work/Rocketeer.jl/Rocketeer.jl/src/lib/rocket.jl:74.

Rocketeer.RocketModule — Method

RocketModule() -> RocketModule

Summary

Empty constructor for a RocketModule.

This uses the default values input_length=5 and n_kernels=100.

Attribution

Programmer

Sasha Petrenko <petrenkos@mst.edu> @AP6YC

Original Authors

Angus Dempster

Francois Petitjean

Geoff Webb

Bibtex Entry

@article{dempster_etal_2020,
    author  = {Dempster, Angus and Petitjean, Francois and Webb, Geoffrey I},
    title   = {ROCKET: Exceptionally fast and accurate time classification using random convolutional kernels},
    year    = {2020},
    journal = {Data Mining and Knowledge Discovery},
    doi     = {https://doi.org/10.1007/s10618-020-00701-z}
}

Citation Links

Papers:

Journal article (DOI)

Preprint

Software

rocket (Python)

Method List / Definition Locations

RocketModule()

defined at /home/runner/work/Rocketeer.jl/Rocketeer.jl/src/lib/rocket.jl:117.

Constants

Rocketeer.ROCKETEER_VERSION — Constant

A constant that contains the version of the installed Rocketeer.jl package.

This value is computed at compile time, so it may be used to programmatically verify the version of OAR that is installed in case a compat entry in your Project.toml is missing or otherwise incorrect.