Developer Index

This page lists the types and functions that are internal to the Rocketeer.jl package. Because they are not part of the public API, these names might change relatively frequently between versions and so should not be relied upon.

All internal names are listed in the Index, and each of these entries link to the docstrings in the Docs section.

"Table" of Contents

Modules	Functions	Types	Constants

Index	Index	Index	Index

Docs	Docs	Docs	Docs

Index

This section contains a list of internal names that link to their corresponding Documentation.

Modules

Functions

Types

Constants

Rocketeer.DOCSTRING_ATTRIBUTION

Rocketeer.ROCKET_DEFAULT_INPUT_LENGTH

Rocketeer.ROCKET_DEFAULT_N_KERNELS

Docs

Documentation for all internal names are listed below.

Modules

Functions

Base.show — Method

show(io::IO, kernel::RocketKernel)

Summary

Overload of the show function for RocketKernel.

Arguments

io::IO: the current IO stream.

kernel::RocketKernel: the RocketKernel to print/display.

Method List / Definition Locations

show(io, kernel)

defined at /home/runner/work/Rocketeer.jl/Rocketeer.jl/src/lib/rocket.jl:223.

Base.show — Method

show(io::IO, rocket::RocketModule)

Summary

Overload of the show function for RocketModule.

Arguments

io::IO: the current IO stream.

rocket::RocketModule: the RocketModule to print/display.

Method List / Definition Locations

show(io, rocket)

defined at /home/runner/work/Rocketeer.jl/Rocketeer.jl/src/lib/rocket.jl:238.

Types

Constants

Rocketeer.DOCSTRING_ATTRIBUTION — Constant

Common docstring: description of attribution for the Rocketeer module for inclusion in relevant docstrings.

Rocketeer.ROCKET_DEFAULT_INPUT_LENGTH — Constant

A constant defining the default input length for constructing RocketModules in the empty constructor.

Rocketeer.ROCKET_DEFAULT_N_KERNELS — Constant

A constant defining the default number of kernels for constructing RocketModules in the empty constructor.